requirements.txt
flask
pytorch
torchvision
